Search System/pyproject.toml
[tool.poetry]
name = "assignmnet3"
version = "1"
description = "poetry setip"
authors = ["group 4"]
license = "(y)"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pinecone-client = "^5.0.1"
boto3 = "^1.35.51"
torch = "^2.5.1"
transformers = "^4.46.1"
python-dotenv = "^1.0.0"
pypdf2 = "^3.0.1"
streamlit = "^1.39.0"
openai = "^1.53.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
